add more code & doc
